Fix autorun macro impersonation
This issue was caused by a recent Roll20 update separating the execution queues of `sendChat()` calls with a callback from those without. MMM impersonation expected a sequence of chat with callback, normal chat, and another chat with callback to result in the same sequence of `on("chat:message")` invocations.

Fixed by introducing a new `$sendChatCallback()` utility that funnels MMM chat callbacks through the normal chat queue.